Update `testenv` mocks
- to only mock the canvas util (which is the one with Jest dependency issues, not the text truncation util itself)
