+++ src/services/canvas/canvas_text_utils.testenv.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export class CanvasTextUtils {
+  constructor(_: any) {}
+
+  computeFontFromElement = (_: HTMLElement) => '';
+
+  textWidth = 0;
+
+  currentText = '';
+  setTextToCheck = (text: string) => {
+    this.currentText = text;
+  };
+}
